clean up and add new tests
